# jmeter-global-orchestrator/api/openapi.yaml
openapi: 3.0.3
info:
  title: jmeter-global-orchestrator
  version: 0.1.0
  summary: Fan-out layer driving many jmeter-local-orchestrators across regions.
  description: |
    The control plane: clients hit this service, and it fans `POST /api/v1/test`
    out to many `jmeter-local-orchestrator` workers, tracking run state in
    Postgres and serving aggregated metrics back from the rollup tables.

    Pod selection is registry-driven and runs claim only IDLE pods belonging
    to their own application. Liveness has two sources: dynamic workers in a
    routed region (`REGIONS=id=url`) are judged by the kubelet through the
    region's `GET /api/v1/workers` (ready → IDLE, dead/absent → LOST with the
    kubelet's reason); operator-declared workers self-register and heartbeat
    every 30 s, and a sweeper marks them LOST after 90 s of silence.
  contact:
    name: Performance Engineering
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8082
    description: Local dev (default HTTP_PORT=8082)
  - url: http://{host}:{port}
    description: Cloud — host and port are deployment-specific
    variables:
      host: { default: global-orchestrator.jmetercloud.local }
      port: { default: "8082" }
...
paths:
...
  /api/v1/applications/{applicationId}/capacity/{region}/pods:
    parameters:
      - { name: applicationId, in: path, required: true, schema: { type: string } }
      - { name: region,        in: path, required: true, schema: { type: string } }
    get:
      tags: [capacity]
      summary: List pods for (application, region) with Ready / In-Use state
      responses:
        "200":
          description: Per-(app, region) capacity snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CapacitySnapshot" }
    post:
      tags: [capacity]
      summary: Spin up one new Ready pod
      description: |
        Allocates the next free `{appName}-{region}-worker-{n}` name,
        creates and starts the container, then waits for the local-orch
        to self-register (typically <30 s; the registry row appears
        asynchronously via POST /registerPod). The response carries the
        allocated podName + the baseUrl the global will use to fan out.
      responses:
        "201":
          description: Spin started. Container created + start command issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpinPodResponse" }
        "404":
          description: applicationId or capacity row missing.
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: |
            Either the spin would exceed `maxAvailable`
            (`APPLICATION_CAPACITY_EXCEEDED`), or this deployment does not
            provision (`PROVISIONING_DISABLED` — see
            `GET /api/v1/platform/capabilities`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CapacityExceededResponse"
                  - $ref: "#/components/schemas/ProvisioningDisabledResponse"
...
  /api/v1/applications/{applicationId}/capacity/{region}/pods/{podName}:
    parameters:
      - { name: applicationId, in: path, required: true, schema: { type: string } }
      - { name: region,        in: path, required: true, schema: { type: string } }
      - { name: podName,       in: path, required: true, schema: { type: string } }
    put:
      tags: [capacity]
      summary: Declare an operator-deployed worker (STATIC provisioning only)
      description: |
        Binds a worker the operator deployed themselves to this
        (application, region). **Static mode only** — `409
        PROVISIONING_REQUIRES_STATIC` when the deployment provisions its own
        workers, where names are allocator-owned and `POST .../pods` (spin)
        is the right call.

        PUT rather than a parallel `/staticPods` collection: the operator
        names the resource (they deployed it and know its name), the call is
        idempotent, and one worker resource keeps the verbs carrying the
        meaning — POST allocates-and-spins, PUT declares an existing one,
        DELETE releases either.

        `podName` MUST equal the `workerId` the worker stamps on its metrics
        or the `metrics."workerMetric"` join breaks.

        The address is probed (`GET {baseUrl}/actuator/health`) before the
        declaration is accepted, so a typo fails here rather than at the next
        run launch. `?force=true` skips that check for a worker that is
        deployed but not up yet.

        `maxAvailable` for the (application, region) is **derived** — it is
        set to the declared worker count on every declare and release. There
        is no separate capacity to approve in this mode, and
        `PUT .../capacity/{region}` is refused accordingly.
      parameters:
        - name: force
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Declare even if the worker does not currently answer.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeclarePodRequest" }
      responses:
        "201":
          description: Worker newly declared.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeclaredPodResponse" }
        "200":
          description: An existing declaration was updated (idempotent re-declare).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeclaredPodResponse" }
        "400":
          description: |
            `INVALID_REQUEST` (malformed podName / baseUrl — non-HTTP scheme,
            embedded credentials, query or fragment) or `WORKER_UNREACHABLE`
            (address did not answer and `force` was not set).
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: applicationId not registered.
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: |
            `PROVISIONING_REQUIRES_STATIC` (this deployment provisions its own
            workers), `POD_BOUND_ELSEWHERE` (already declared to another
            application), or `POD_IN_USE` (re-addressing a worker held by an
            active run would break that run's follow-up calls to it).
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    delete:
      tags: [capacity]
      summary: Release a pod — drain (dynamic) or undeclare (static)
      description: |
        Under `PROVISIONING_MODE=DYNAMIC` this is a full **drain**: the
        container is stopped and removed, then the registry row is deleted.
        Under `STATIC` it is an **undeclare** — the registry row is deleted
        and the operator's worker keeps running, because the control plane
        does not own it. The in-use guard, the stale-binding release and the
        response shape are identical in both modes; only `containerStopped`
        differs.

        Refuses 409 `POD_IN_USE` when the pod is held by an active run whose
        worker **container is still running** — a genuinely busy pod. To drain
        it, abort the run first (`POST /api/v1/runs/{runId}/abort`).

        STALE-BINDING tolerance: if the pod is bound to a non-terminal run but
        its container is no longer running (a zombie run — the worker died and
        no heartbeat flipped the run terminal), the binding is treated as stale.
        The drain proceeds, releasing the dead member binding so a re-spun
        same-name worker won't re-bind to the zombie run; the response carries
        `staleBindingReleased: true`. (The zombie run itself stays non-terminal
        until you abort it — drain frees the pod, abort ends the run.)
      responses:
        "200":
          description: Drained.
          content:
            application/json:
              schema:
                type: object
                properties:
                  podName: { type: string }
                  drained: { type: boolean }
                  containerStopped:
                    type: boolean
                    description: |
                      True when the control plane stopped and removed the
                      worker container (dynamic mode). False in static mode —
                      the registry row was removed and the operator's worker
                      was left running.
                  staleBindingReleased:
                    type: boolean
                    description: |
                      Present + true only when the drain released a stale
                      binding to a non-terminal run whose container was gone.
        "404":
          description: Pod not bound to (applicationId, region).
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Pod is held by an active run whose container is still running.
          content: { application/json: { schema: { $ref: "#/components/schemas/PodInUseResponse" } } }

  # ────────────────────────────────────────────────────────────────────────
  # Platform capabilities
  # ────────────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    CapacitySnapshot:
      type: object
      required: [applicationId, region, maxAvailable, provisioned, ready, inUse, spinnable, pods]
      properties:
        applicationId: { type: string }
        region:        { type: string }
        maxAvailable:  { type: integer, description: "Operator-set ceiling from applicationCapacity.maxAvailable." }
        provisioned:   { type: integer, description: "ready + inUse — pods that exist as containers right now." }
        ready:         { type: integer, description: "Pods running + idle (not currently held by any run)." }
        inUse:         { type: integer, description: "Pods currently held by an active runFleetMember." }
        spinnable:     { type: integer, description: "maxAvailable - provisioned, clamped at 0. What POST /pods can create right now." }
        pods:
          type: array
          items: { $ref: "#/components/schemas/PodView" }
    PodView:
      type: object
      required: [podName, state, containerRunning, runsServed]
      properties:
        podName:          { type: string, example: "payments-us-east-worker-1" }
        state:            { type: string, enum: [READY, IN_USE, LOST, STARTING, UNKNOWN, RECYCLING],
                            description: "RECYCLING is reported when the pod row is in DRAINING_FOR_RECYCLE state (WORKER-HYGIENE Phase D)." }
        containerRunning: { type: boolean, description: "Live readback from `docker inspect` (provisioner.isRunning). False on daemon outages." }
        lastHeartbeat:    { type: string, format: date-time, nullable: true }
        blockedBy:        { $ref: "#/components/schemas/PodBlockedBy", nullable: true }
        runsServed:       { type: integer, format: int64, description: "WORKER-HYGIENE Phase F1 — count of runs claimed against this pod." }
        imageDigest:      { type: string, nullable: true, description: "WORKER-HYGIENE Phase F1 — sha256 ID of the image; null for legacy rows." }
        provisionedAt:    { type: string, format: date-time, nullable: true, description: "WORKER-HYGIENE Phase F1 — container-create timestamp; null for legacy rows." }
    PodBlockedBy:
      type: object
      description: "Set on a PodView when state=IN_USE. Surfaces the run holding the pod so the UI can render 'cannot drain — running test {runId}'."
      properties:
        runId:       { type: string }
        state:       { type: string, description: "Run state — PREPARING / STARTING / RUNNING / DRAINING." }
        startedAt:   { type: string, format: date-time, nullable: true }
        initiatedBy: { type: string }
    SpinPodResponse:
      type: object
      required: [podName, applicationId, region, baseUrl, provisioned, maxAvailable]
      properties:
        podName:       { type: string, example: "payments-us-east-worker-1" }
        applicationId: { type: string }
        region:        { type: string }
        baseUrl:       { type: string, example: "http://payments-us-east-worker-1:8080" }
        provisioned:   { type: integer, description: "Provisioned count after this spin." }
        maxAvailable:  { type: integer }
        ready:
          type: boolean
          description: "True when the kubelet reported the pod ready within the wait; false means it is still STARTING and becomes READY through the liveness probe."
...
    DeclarePodRequest:
      type: object
      required: [baseUrl]
      properties:
        baseUrl:
          type: string
          maxLength: 512
          description: |
            Absolute http(s) address the control plane can reach this worker
            at, e.g. `http://payments-na-east-worker-1.workers:8080`. No
            embedded credentials, query string or fragment. A trailing slash
            is normalised away.
          example: http://payments-na-east-worker-1.workers:8080
    DeclaredPodResponse:
      type: object
      required: [podName, applicationId, region, baseUrl, source, reachable, declared, maxAvailable]
      properties:
        podName:       { type: string }
        applicationId: { type: string }
        region:        { type: string }
        baseUrl:       { type: string }
        source:        { type: string, enum: [STATIC] }
        reachable:
          type: boolean
          description: |
            Whether the worker answered its health endpoint at declare time.
            False only when `?force=true` was used.
        declared:
          type: integer
          description: Declared workers for this (application, region) after the call.
        maxAvailable:
          type: integer
          description: |
            The derived capacity written for this (application, region) —
            always equal to `declared` in static mode.
    ProvisioningDisabledResponse:
      type: object
      required: [code, message, action, provisioningMode]
      properties:
        code:             { type: string, enum: [PROVISIONING_DISABLED] }
        message:          { type: string }
        action:           { type: string, description: The refused operation, e.g. "spin a worker". }
        provisioningMode: { type: string, enum: [STATIC] }
    CapacityExceededResponse:
      type: object
      required: [code, message, provisioned, maxAvailable]
      properties:
        code:         { type: string, enum: [APPLICATION_CAPACITY_EXCEEDED] }
        message:      { type: string }
        provisioned:  { type: integer }
        maxAvailable: { type: integer }
...
    PodInUseResponse:
      type: object
      required: [code, message, podName, blockedBy]
      properties:
        code:      { type: string, enum: [POD_IN_USE] }
        message:   { type: string }
        podName:   { type: string }
        blockedBy: { $ref: "#/components/schemas/PodBlockedBy" }
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - REGION_UNREACHABLE
            - REGIONAL_CALL_FAILED
            - RUN_NOT_FOUND
            - INVALID_REQUEST
            - INTERNAL_ERROR
            - INSUFFICIENT_CAPACITY
            - POD_NOT_REGISTERED
            - UNKNOWN_REGION
            - FLEET_SIZE_EXCEEDED
            # Phase 3 capacity rework error codes — surfaced by CapacityController.
            - APPLICATION_NOT_FOUND
            - CAPACITY_REGION_NOT_FOUND
            - APPLICATION_CAPACITY_EXCEEDED
            - CAPACITY_SHRINK_BELOW_PROVISIONED
            - POD_NOT_FOUND
            - POD_IN_USE
            # AI analysis error codes — surfaced by AiController.
            - AI_DISABLED
            - AI_QUOTA_EXCEEDED
            - AI_UPSTREAM_ERROR
        message: { type: string }
        shortfall:
          type: array
          description: |
            Present on `INSUFFICIENT_CAPACITY` responses — per-region
            breakdown of requested-vs-claimed counts so the UI can
            highlight which regions fell short.
          items:
            type: object
            properties:
              region:    { type: string }
              requested: { type: integer }
              claimed:   { type: integer }
        regions:
          type: array
          description: Present on `UNKNOWN_REGION` — the region names the caller specified that the registry has never seen.
          items: { type: string }
        requested:
          type: integer
          description: Present on `FLEET_SIZE_EXCEEDED` — the total allocation count the caller submitted.
        max:
          type: integer
          description: Present on `FLEET_SIZE_EXCEEDED` — the per-run fleet-size cap.

    # ──────────────────────────────────────────────────────────────────────
    # AI analysis
    # ──────────────────────────────────────────────────────────────────────
...